+++ src/modules/all.less
@@ -16,4 +16,4 @@
 
 @import "./esl-alert/core.less";
 
-@import "./forms/all.less";
+@import "./esl-forms/all.less";
