+++ src/modules/esl-forms/esl-select/core/esl-select.less
@@ -1,11 +1,11 @@
-esl-select {
+.esl-select {
   display: inline-block;
   position: relative;
 
   min-width: 300px;
   max-width: 100%;
 
-  &.esl-select [esl-select-target] {
+  & [esl-select-target] {
     /* Visually hidden */
     position: absolute;
     width: 0;
